# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/organizations/me": {
      "get": {
        "description": "Retrieve a list of organizations associated with a specific user using cursor-based pagination",
        "operationId": "OrganizationController_getOrganizationsByOwner",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items to return",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Number of items to skip",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortDirection",
            "required": false,
            "in": "query",
            "description": "Sort direction (asc or desc)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor for pagination",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "Keyword to search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CompanyDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Get organizations by user ID with cursor pagination",
        "tags": [
          "Organization"
        ]
      }
    },
    "/organizations/check-name/{name}": {
      "get": {
        "description": "Check if a company name exists",
        "operationId": "OrganizationController_checkNameExists",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CheckOrganizationNameResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Check if a company name exists",
        "tags": [
          "Organization"
        ]
      }
    },
    "/organizations/{orgId}": {
      "get": {
        "description": "Retrieve an organization by its ID",
        "operationId": "OrganizationController_getOrganization",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/GetCompanyDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Get organization by ID",
        "tags": [
          "Organization"
        ]
      },
      "delete": {
        "description": "Delete an organization. Requires confirmation by entering the exact organization name.",
        "operationId": "OrganizationController_deleteOrganization",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/String"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Delete an organization",
        "tags": [
          "Organization"
        ]
      }
    },
    "/organizations": {
      "post": {
        "description": "Create a new organization",
        "operationId": "OrganizationController_createOrganization",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/String"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Create a new organization",
        "tags": [
          "Organization"
        ]
      }
    },
    "/organizations/{orgId}/basic-info": {
      "patch": {
        "description": "Update basic information including name, description, websiteUrl, and phone. Only sends changed fields.",
        "operationId": "OrganizationController_updateOrganizationBasicInfo",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationBasicInfoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OrganizationWithDetailsDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Update organization basic information",
        "tags": [
          "Organization"
        ]
      }
    },
    "/organizations/{orgId}/locations": {
      "patch": {
        "description": "Replace all organization locations with new list. Frontend sends FULL list of current locations. All existing locations will be deleted and replaced with new ones.",
        "operationId": "OrganizationController_updateOrganizationLocations",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationLocationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/String"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Update organization locations",
        "tags": [
          "Organization"
        ]
      }
    },
    "/organizations/{orgId}/additional-info": {
      "patch": {
        "description": "Update culture and benefits information. Only for COMPANY type organizations. Only sends changed fields.",
        "operationId": "OrganizationController_updateOrganizationAdditionalInfo",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationAdditionalInfoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OrganizationWithDetailsDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Update organization additional information (Culture & Benefits)",
        "tags": [
          "Organization"
        ]
      }
    },
    "/organizations/{orgId}/email": {
      "patch": {
        "description": "Request to update organization email. If email is NOT verified, it will be updated immediately. If email IS verified, an OTP will be sent to the NEW email for confirmation.",
        "operationId": "OrganizationController_updateOrganizationEmail",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/String"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Update organization email (Request)",
        "tags": [
          "Organization"
        ]
      }
    },
    "/organizations/{orgId}/email/confirm": {
      "post": {
        "description": "Confirm organization email change using the OTP code sent to the NEW email address. This will update the email and reset verifiedAt to null.",
        "operationId": "OrganizationController_confirmUpdateOrganizationEmail",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmUpdateOrganizationEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/String"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Confirm email change with OTP",
        "tags": [
          "Organization"
        ]
      }
    },
    "/organizations/{orgId}/email/send-verification": {
      "post": {
        "description": "Send a 6-digit OTP code to the organization's email for verification. OTP expires in 10 minutes.",
        "operationId": "OrganizationController_sendEmailVerificationOtp",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendEmailVerificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/String"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Send email verification OTP",
        "tags": [
          "Organization"
        ]
      }
    },
    "/organizations/{orgId}/email/verify": {
      "post": {
        "description": "Verify the organization's email using the OTP code sent via email. Sets verifiedAt timestamp upon successful verification.",
        "operationId": "OrganizationController_verifyOrganizationEmail",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOrganizationEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/String"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Verify organization email with OTP",
        "tags": [
          "Organization"
        ]
      }
    },
    "/organizations/all": {
      "get": {
        "description": "Get all organizations (only basic information)",
        "operationId": "OrganizationController_getAllOrganizations",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items to return",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Number of items to skip",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortDirection",
            "required": false,
            "in": "query",
            "description": "Sort direction (asc or desc)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor for pagination",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "Keyword to search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/String"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Get all organizations",
        "tags": [
          "Organization"
        ]
      }
    },
    "/organizations/{orgId}/users-to-invite": {
      "get": {
        "description": "Retrieve a list of users who can be invited to join a specific organization",
        "operationId": "OrganizationController_getUsersToInvite",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items to return",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Number of items to skip",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortDirection",
            "required": false,
            "in": "query",
            "description": "Sort direction (asc or desc)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor for pagination",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "Keyword to search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get available users to invite to an organization",
        "tags": [
          "Organization"
        ]
      }
    },
    "/organizations/{orgId}/logo": {
      "patch": {
        "description": "Upload a new logo for the organization. Accepts image files (JPEG, PNG, WebP). Max size: 5MB",
        "operationId": "OrganizationController_updateOrganizationLogo",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/String"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Update organization logo",
        "tags": [
          "Organization"
        ]
      }
    },
...
  },
  "info": {
    "title": "AI Recruit Documentation",
    "description": "This is the API Docs for using internally.",
    "version": "1.0",
    "contact": {},
    "termsOfService": "http://swagger.io/terms/",
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local server 1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "SUCCESS"
          },
          "message": {
            "type": "string",
            "example": "Request was successful."
          },
          "data": {
            "type": "object"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "CheckOrganizationNameResponseDto": {
        "type": "object",
        "properties": {
          "exists": {
            "type": "boolean"
          }
        },
        "required": [
          "exists"
        ]
      },
...
      "String": {
        "type": "object",
        "properties": {}
      },
...
      "UpdateOrganizationLocationDto": {
        "type": "object",
        "properties": {
          "locations": {
            "description": "List of organization locations",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateLocationDto"
            }
          }
        },
        "required": [
          "locations"
        ]
      },
...
      "DeleteOrganizationDto": {
        "type": "object",
        "properties": {
          "confirmationName": {
            "type": "string",
            "example": "My Company Name",
            "description": "Organization name for confirmation. Must match exactly to proceed with deletion."
          }
        },
        "required": [
          "confirmationName"
        ]
      },
...
    }
  },
...
}